% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ConstructLocTree.R
\name{ConstructLocTree}
\alias{ConstructLocTree}
\title{Connect and parse UniProt information.}
\usage{
ConstructLocTree(ProteinDataObject,directorypath = NULL)
}
\arguments{
\item{ProteinDataObject}{input a Dataframe returned from GetNamesTaxa function}

\item{directorypath}{path to save txt file containig results returened by the function.}
}
\description{
This Function is used to plot location's Tree in the data of the accession/s in the chromosomes.
}
\author{
Mohmed Soudy \email{Mohamed.soudy@57357.com} and Ali Mostafa \email{ali.mo.anwar@std.agr.cu.edu.eg}
}
