% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ConstructPhylogeny.R
\name{ConstructPhylogeny}
\alias{ConstructPhylogeny}
\title{Connect and parse UniProt information.}
\usage{
ConstructPhylogeny(ProteinDataObject , directorypath = NULL)
}
\arguments{
\item{ProteinDataObject}{Dataframe returned from GetSequence function.}

\item{directorypath}{path to save files containig results returened by the function ( default = NA ).}
}
\description{
This Function is used to construct phylogenetic tree from protein sequence applying msa on the sequence list.
}
\note{
To use this package you MUst INSTALL msa package from Bioconductor ,if no dir_path was given ( default = NA ) the function will only view the plot and will not save it
}
\author{
Mohmed Soudy \email{Mohamed.soudy@57357.com} and Ali Mostafa \email{ali.mo.anwar@std.agr.cu.edu.eg}
}
