% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PathwayAnalysis.R
\name{Enrichment.CC}
\alias{Enrichment.CC}
\title{Connect and parse UniProt information}
\usage{
Enrichment.CC(Accs,OS="hsapiens",p_value=0.05,directorypath=NULL,top=10)
}
\arguments{
\item{Accs}{Vector of UniProt Accession/s or genes}

\item{OS}{organism name Example: human - 'hsapiens', mouse - 'mmusculus'}

\item{p_value}{custom p-value threshold for significance, default = 0.05}

\item{directorypath}{Path to save output plot}

\item{top}{Top N terms to be visualized}
}
\description{
This function is used for Enrichment analysis of cellular component of given list of genes or proteins
}
\author{
Mohmed Soudy \email{Mohamed.soudy@57357.com} and Ali Mostafa \email{ali.mo.anwar@std.agr.cu.edu.eg}
}
