% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PathwayAnalysis.R
\name{Enrichment.gprofiler}
\alias{Enrichment.gprofiler}
\title{Connect and parse UniProt information}
\usage{
Enrichment.gprofiler(Accs, sources, OS = "hsapiens", p_value = 0.05, directorypath = NULL)
}
\arguments{
\item{Accs}{Vector of UniProt Accession/s or genes}

\item{sources}{a vector of data sources to use, these include GO (GO:BP, GO:MF, GO:CC to select a particular GO branch), KEGG, REAC}

\item{OS}{organism name Example: human - 'hsapiens', mouse - 'mmusculus'}

\item{p_value}{custom p-value threshold for significance, default = 0.05}

\item{directorypath}{path to save excel file and plot returened by the function.}
}
\description{
This function is used for Enrichment analysis of given list of genes or proteins
}
\author{
Mohmed Soudy \email{Mohamed.soudy@57357.com} and Ali Mostafa \email{ali.mo.anwar@std.agr.cu.edu.eg}
}
