% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PathwayAnalysis.R
\name{Enrichment.React}
\alias{Enrichment.React}
\title{Connect and parse UniProt information}
\usage{
Enrichment.React(AccList, organism = "human", pvalueCutoff = 0.05, directorypath = NULL)
}
\arguments{
\item{AccList}{Vector of UniProt Accession/s or genes}

\item{organism}{one of "human", "rat", "mouse", "celegans", "yeast", "zebrafish", "fly"}

\item{pvalueCutoff}{Cutoff value of pvalue, default = 0.05}

\item{directorypath}{path to save excel file and plot returened by the function}
}
\description{
This function is used for Enrichment analysis of given list of genes or proteins based on Reactome API
}
\author{
Mohmed Soudy \email{Mohamed.soudy@57357.com} and Ali Mostafa \email{ali.mo.anwar@std.agr.cu.edu.eg}
}
