% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Getdiseases.R
\name{Get.diseases}
\alias{Get.diseases}
\title{Connect and parse UniProt information.}
\usage{
Get.diseases(Pathology_object , directorypath = NULL)
}
\arguments{
\item{Pathology_object}{Dataframe retrieved from UniprotR Function "GetPathology_Biotech"}

\item{directorypath}{path to save Output file}
}
\description{
This Function is used to get diseases associated with proteins.
}
\author{
Mohmed Soudy \email{Mohamed.soudy@57357.com} and Ali Mostafa \email{ali.mo.anwar@std.agr.cu.edu.eg}
}
