% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetAccessionList.R
\name{GetAccessionList}
\alias{GetAccessionList}
\title{Connect and parse UniProt information.}
\usage{
GetAccessionList(DataObjPath)
}
\arguments{
\item{DataObjPath}{input path of excel file}
}
\value{
a vector of UniProt Accession/s
}
\description{
This function can be used to get a list of UniProt Accession/s from a csv file.
}
\author{
Mohmed Soudy \email{Mohamed.soudy@57357.com} and Ali Mostafa \email{ali.mo.anwar@std.agr.cu.edu.eg}
}
