% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetFamily_Domains.R
\name{GetFamily_Domains}
\alias{GetFamily_Domains}
\title{Connect and parse UniProt Family Domains information.}
\usage{
GetFamily_Domains(ProteinAccList , directorypath = NULL)
}
\arguments{
\item{ProteinAccList}{Vector of UniProt Accession/s}

\item{directorypath}{path to save excel file containig results returened by the function.}
}
\value{
DataFrame where rows names are the accession
     and columns contains the information retrieved from the UniProt
}
\description{
The function is work to retrieve Family Domains data from UniProt for a
list of proteins accessions.For more information about what included
in the Family Domains data see https://www.uniprot.org/help/return_fields.
}
\note{
The function also, Creates a csv file with the retrieved information.
}
\author{
Mohmed Soudy \email{Mohamed.soudy@57357.com} and Ali Mostafa \email{ali.mo.anwar@std.agr.cu.edu.eg}
}
