% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetMiscellaneous.R
\name{GetMiscellaneous}
\alias{GetMiscellaneous}
\title{Connect and parse UniProt Miscellaneous information.}
\usage{
GetMiscellaneous(ProteinAccList , directorypath = NULL)
}
\arguments{
\item{ProteinAccList}{Vector of UniProt Accession/s}

\item{directorypath}{path to save excel file containig results returened by the function.}
}
\value{
DataFrame where rows names are the accession
     and columns contains the information retrieved from the UniProt
}
\description{
The function is work to retrieve Miscellaneous data from UniProt for a list
of proteins accessions.For more information about what included in the
Miscellaneous data see https://www.uniprot.org/help/return_fields.
}
\note{
The function also, Creates a csv file with the retrieved information.
}
\examples{
Obj <- GetMiscellaneous("O14520")

}
\author{
Mohmed Soudy \email{Mohamed.soudy@57357.com} and Ali Mostafa \email{ali.mo.anwar@std.agr.cu.edu.eg}
}
