% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetPathology_Biotech.R
\name{GetPathology_Biotech}
\alias{GetPathology_Biotech}
\title{Connect and parse UniProt Pathology_Biotech information.}
\usage{
GetPathology_Biotech(ProteinAccList , directorypath = NULL)
}
\arguments{
\item{ProteinAccList}{Vector of UniProt Accession/s}

\item{directorypath}{path to save excel file containig results returened by the function.}
}
\value{
DataFrame where rows names are the accession
     and columns contains the information retrieved from the UniProt
}
\description{
The function is work to retrieve Pathology_Biotech data from UniProt for
a list of proteins accessions.For more information about what included in the
Pathology_Biotech data see https://www.uniprot.org/help/uniprotkb_column_names.
}
\note{
The function also, Creates a csv file with the retrieved information.
}
\examples{
Obj <- GetPathology_Biotech("O14520")

}
\author{
Mohmed Soudy \email{Mohamed.soudy@57357.com} and Ali Mostafa \email{ali.mo.anwar@std.agr.cu.edu.eg}
}
