% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetpdbStructure.R
\name{GetpdbStructure}
\alias{GetpdbStructure}
\title{Connect and parse UniProt information.}
\usage{
GetpdbStructure(ProteinAccList , directorypath = NULL)
}
\arguments{
\item{ProteinAccList}{input a vector of UniProt Accession/s}

\item{directorypath}{path to save excel file containig results returened by the function.}
}
\description{
The function is work to retrieving GetpdbStructure and download it
to user directory.
}
\author{
Mohmed Soudy \email{Mohamed.soudy@57357.com} and Ali Mostafa \email{ali.mo.anwar@std.agr.cu.edu.eg}
}
