% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetProteoAnnotator.R
\name{GetProteinAnnontate}
\alias{GetProteinAnnontate}
\title{Connect and parse UniProt information.}
\usage{
GetProteinAnnontate(ProteinAccList , columns)
}
\arguments{
\item{ProteinAccList}{a vector of UniProt Accession/s}

\item{columns}{a vector of UniProtKB column names}
}
\value{
DataFrame where rows names are the accession
     and columns contains the information retrieved from the UniProt
}
\description{
The function is work to retrieve user-defined information data from UniProt
for a list of proteins accessions For more
information see https://www.uniprot.org/help/uniprotkb_column_names
}
\examples{
Obj <- GetProteinInteractions("O14520")

}
\author{
Mohmed Soudy \email{Mohamed.soudy@57357.com} and Ali Mostafa \email{ali.mo.anwar@std.agr.cu.edu.eg}
}
