% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetProteinFunction.R
\name{GetProteinFunction}
\alias{GetProteinFunction}
\title{Connect and parse UniProt Protein Function information.}
\usage{
GetProteinFunction(ProteinAccList , directorypath = NULL)
}
\arguments{
\item{ProteinAccList}{Vector of UniProt Accession/s}

\item{directorypath}{path to save excel file containig results returened by the function.}
}
\value{
DataFrame where rows names are the accession
     and columns contains the information of protein function roles from the UniProt
}
\description{
The function is work to retrieve Protein Function data from UniProt for
a list of proteins accessions.For more information about what included in the
Protein Function data see https://www.uniprot.org/help/return_fields.
}
\note{
The function also, Creates a csv file with the retrieved information.
}
\examples{
Obj <- GetProteinFunction("O14520")

}
\author{
Mohmed Soudy \email{Mohamed.soudy@57357.com} and Ali Mostafa \email{ali.mo.anwar@std.agr.cu.edu.eg}
}
