% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetproteinNetwork.R
\name{GetproteinNetwork}
\alias{GetproteinNetwork}
\title{Connect and parse stringdb information.}
\usage{
GetproteinNetwork(ProteinAccList , directorypath = NULL)

GetproteinNetwork(ProteinAccList , directorypath = NULL)
}
\arguments{
\item{ProteinAccList}{input a vector of UniProt Accession/s}

\item{directorypath}{path to save excel file containig results returened by the function.}
}
\description{
This function is connecting to stringdb and retrieve all possible interactions
for the searched protein/s.
}
\author{
Mohmed Soudy \email{Mohamed.soudy@57357.com} and Ali Mostafa \email{ali.mo.anwar@std.agr.cu.edu.eg}
}
