% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetProteomeFasta.R
\name{GetProteomeFasta}
\alias{GetProteomeFasta}
\title{Connect and DOWNLOAD Proteome info.}
\usage{
GetProteomeFasta(ProteomeID , directorypath = NULL)
}
\arguments{
\item{ProteomeID}{Proteome ID from UniProt}

\item{directorypath}{path to save FASTA file containig results returened by the function.}
}
\description{
The function is work to retrieve proteome information in FASTA format based on proteome id.
}
\note{
The function Download fasta format of proteome.
}
\author{
Mohmed Soudy \email{Mohamed.soudy@57357.com} and Ali Mostafa \email{ali.mo.anwar@std.agr.cu.edu.eg}
}
