% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetProteomeInfo.R
\name{GetProteomeInfo}
\alias{GetProteomeInfo}
\title{Connect and DOWNLOAD Proteome info.}
\usage{
GetProteomeInfo(ProteomeID , directorypath = NULL)
}
\arguments{
\item{ProteomeID}{Proteome ID from UniProt}

\item{directorypath}{path to save CSV file containig results returened by the function.}
}
\description{
The function is work to retrieve proteome information based on proteome id.
}
\note{
The function Download csv Info of proteome.
}
\author{
Mohmed Soudy \email{Mohamed.soudy@57357.com} and Ali Mostafa \email{ali.mo.anwar@std.agr.cu.edu.eg}
}
