% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetPTM_Processing.R
\name{GetPTM_Processing}
\alias{GetPTM_Processing}
\title{Connect and parse UniProt PTM_Processsing information.}
\usage{
GetPTM_Processing(ProteinAccList, directorypath = NULL)
}
\arguments{
\item{ProteinAccList}{Vector of UniProt Accession/s}

\item{directorypath}{path to save excel file containig results returened by the function}
}
\value{
DataFrame where rows names are the accession
     and columns contains the information retrieved from the UniProt
}
\description{
The function is work to retrieve PTM_Processsing data from UniProt for a list of proteins accessions.
For more information about what included in the PTM_Processsing data
see https://www.uniprot.org/help/return_fields.
}
\note{
The function also, Creates a csv file with the retrieved information.
}
\examples{
Obj <- GetPTM_Processing("O14520" )

}
\author{
Mohmed Soudy \email{Mohamed.soudy@57357.com} and Ali Mostafa \email{ali.mo.anwar@std.agr.cu.edu.eg}
}
