% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GETSeqFastaUniprot.R
\name{GETSeqFastaUniprot}
\alias{GETSeqFastaUniprot}
\title{Connect and parse UniProt information.}
\usage{
GETSeqFastaUniprot(Accessions,FilePath = NULL, FileName = NULL)
}
\arguments{
\item{Accessions}{Vector of UniProt Accession/s}

\item{FilePath}{path of directory to save the output fasta.}

\item{FileName}{Name of the fasta file.}
}
\description{
This Function is used to get Sequence information of accession/s from Uniprot as a Fasta file.
}
\author{
Mohmed Soudy \email{Mohamed.soudy@57357.com} and Ali Mostafa \email{ali.mo.anwar@std.agr.cu.edu.eg}
}
