% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetSeqLength.R
\name{GetSeqLength}
\alias{GetSeqLength}
\title{Connect and parse UniProt Sequences information.}
\usage{
GetSeqLength(ProteinAccList, directorypath = NULL)
}
\arguments{
\item{ProteinAccList}{Vector of UniProt Accession/s}

\item{directorypath}{path to save excel file containig results returened by the function}
}
\value{
DataFrame where rows names are the accession
     and columns contains the information retrieved from the UniProt
}
\description{
The function is work to retrieve Sequence's Length data from Uniparc for a list of proteins accessions.
This function was added to overcome the NAs returned when Uniprot database deleted the protein from the database
}
\note{
The function also, Creates a csv file with the retrieved information.
}
\examples{
Obj <- GetSequences("O14520")

}
\author{
Mohmed Soudy \email{Mohamed.soudy@57357.com} and Ali Mostafa \email{ali.mo.anwar@std.agr.cu.edu.eg}
}
