% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetSequenceIso.R
\name{GetSequenceIso}
\alias{GetSequenceIso}
\title{Connect and parse UniProt Sequences information.}
\usage{
GetSequenceIso(ProteinAccList, directorypath = NULL)
}
\arguments{
\item{ProteinAccList}{Vector of UniProt Accession/s}

\item{directorypath}{path to save excel file containig results returened by the function}
}
\value{
DataFrame where rows names are the accession
     and columns contains the information retrieved from the UniProt
}
\description{
The function is work to retrieve protein's Sequence data from Uniparc for a list of proteins accessions.
This function was added to handle isoformes
}
\note{
The function also, Creates a csv file with the retrieved information.
}
\examples{
Obj <- GetSequenceIso("O14520")

}
\author{
Mohmed Soudy \email{Mohamed.soudy@57357.com} and Ali Mostafa \email{ali.mo.anwar@std.agr.cu.edu.eg}
}
