% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetSubcellular_location.R
\name{GetSubcellular_location}
\alias{GetSubcellular_location}
\title{Connect and parse UniProt protein Subcellular location information.}
\usage{
GetSubcellular_location(ProteinAccList, directorypath = NULL)
}
\arguments{
\item{ProteinAccList}{Vector of UniProt Accession/s.}

\item{directorypath}{path to save excel file containig results returened by the function.}
}
\value{
DataFrame where rows names are the accession
     and columns contains the information about Subcellular location of protein from the UniProt
}
\description{
The function is work to retrieve protein Subcellular location data from UniProt for a list of proteins accessions.
For more information about what included in the protein Subcellular location
data see https://www.uniprot.org/help/return_fields.
}
\note{
The function also, Creates a csv file with the retrieved information.
}
\author{
Mohmed Soudy \email{Mohamed.soudy@57357.com} and Ali Mostafa \email{ali.mo.anwar@std.agr.cu.edu.eg}
}
