% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetTaxaidentifier.R
\name{GetTaxaidentifier}
\alias{GetTaxaidentifier}
\title{Connect and parse UniProt protein taxonomic information.}
\usage{
GetTaxaidentifier(ProteinAccList, directorypath = NULL)
}
\arguments{
\item{ProteinAccList}{Vector of UniProt Accession/s}

\item{directorypath}{path to save excel file containig results returened by the function.}
}
\value{
DataFrame where rows names are the accession
     and columns contains the Taxonomic Information of protein from the UniProt
}
\description{
The function is work to retrieve Miscellaneous data from UniProt for a list of proteins accessions.
For more information about what included in the Miscellaneous
data see https://www.uniprot.org/help/uniprotkb_column_names.
}
\note{
The function also, Creates a csv file with the retrieved information.
}
\examples{
Obj <- GetTaxaidentifier("O14520")

}
\author{
Mohmed Soudy \email{Mohamed.soudy@57357.com} and Ali Mostafa \email{ali.mo.anwar@std.agr.cu.edu.eg}
}
