% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotGoInfo.R
\name{Goparse}
\alias{Goparse}
\title{Connect and parse UniProt information.}
\usage{
Goparse(GOObj , index = 3)
}
\arguments{
\item{GOObj}{Dataframe.}

\item{index}{idex of Go term in GoObj}
}
\description{
This Function is used to parse data retrieved from UniprotR Function "GetProteinGOInfo".
}
\author{
Mohmed Soudy \email{Mohamed.soudy@57357.com} and Ali Mostafa \email{ali.mo.anwar@std.agr.cu.edu.eg}
}
