% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HandleBadRequests.R
\name{HandleBadRequests}
\alias{HandleBadRequests}
\title{Handle bad requests
This Function is used to handle possible errors when trying to get url response.}
\usage{
HandleBadRequests(RequestCode)
}
\arguments{
\item{RequestCode}{Response returned from url}
}
\value{
None
}
\description{
Handle bad requests
This Function is used to handle possible errors when trying to get url response.
}
\examples{
HandleBadRequests(400)

}
\author{
Mohmed Soudy \email{Mohamed.soudy@57357.com} and Ali Mostafa \email{ali.mo.anwar@std.agr.cu.edu.eg}
}
