% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotAcidity.R
\name{PlotAcidity}
\alias{PlotAcidity}
\title{Connect and parse UniProt information.}
\usage{
PlotAcidity(SeqDataObjPath , directorypath = NULL)
}
\arguments{
\item{SeqDataObjPath}{Dataframe retrieved from UniprotR Function "GetSequences"}

\item{directorypath}{path to save Output plot.}
}
\description{
This Function is used to plot proteins acidity retrieved from "GetSequences" Function.
}
\author{
Mohmed Soudy \email{Mohamed.soudy@57357.com} and Ali Mostafa \email{ali.mo.anwar@std.agr.cu.edu.eg}
}
