% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotEnrichedGO.R
\name{PlotEnrichedGO}
\alias{PlotEnrichedGO}
\title{Connect and parse UniProt information}
\usage{
PlotEnrichedGO(Accs,OS="hsapiens",p_value=0.05,Path=NULL,theme="aaas",width=7,height=7)
}
\arguments{
\item{Accs}{Vector of UniProt Accession/s or genes}

\item{OS}{organism name Example: human - 'hsapiens', mouse - 'mmusculus'}

\item{p_value}{custom p-value threshold for significance, default = 0.05}

\item{Path}{Path to save output plot}

\item{theme}{optional parameter to generate specific theme for journals ex: "aaas", "nature", "lancet", "jama"}

\item{width}{width of the generated plot}

\item{height}{height of the generated plot}
}
\description{
This function is used to generate a combined plot for the enriched Gene Ontology terms
}
\author{
Mohmed Soudy \email{Mohamed.soudy@57357.com} and Ali Mostafa \email{ali.mo.anwar@std.agr.cu.edu.eg}
}
