% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotGenesNetwork.R
\name{PlotGenesNetwork}
\alias{PlotGenesNetwork}
\title{Connect and parse UniProt information.}
\usage{
PlotGenesNetwork(ProteinDataObject , directorypath = NULL)
}
\arguments{
\item{ProteinDataObject}{Dataframe retrieved from UniprotR Function "GetNamesTaxa"}

\item{directorypath}{path to save Output plot.}
}
\description{
This Function is used to cluster proteins based on primary genes retrieved from "GetNamesTaxa" Function.
}
\author{
Mohmed Soudy \email{Mohamed.soudy@57357.com} and Ali Mostafa \email{ali.mo.anwar@std.agr.cu.edu.eg}
}
