% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotGOAll.R
\name{PlotGOAll}
\alias{PlotGOAll}
\title{Connect and parse UniProt information.}
\usage{
PlotGOAll(GOObj, Top = 10, directorypath = NULL, width = width, height = height)
}
\arguments{
\item{GOObj}{Dataframe returned from UniprotR Function "GetProteinGOInfo"}

\item{Top}{Number of molecular functions to be visualized}

\item{directorypath}{path to save Output plot.}

\item{width}{width of the generated plot}

\item{height}{height of the generated plot}
}
\description{
This Function is used to plot the retrieved Gene Ontology from function 'GetProteinGOInfo'.
}
\author{
Mohmed Soudy \email{Mohamed.soudy@57357.com} and Ali Mostafa \email{ali.mo.anwar@std.agr.cu.edu.eg}
}
