% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotGoannotation.R
\name{plotGoannotation}
\alias{plotGoannotation}
\title{Connect and parse UniProt information.}
\usage{
plotGoannotation(ProteinDataObject,directorypath = NULL)
}
\arguments{
\item{ProteinDataObject}{input a Dataframe returned from GetProteinGOInfo function}

\item{directorypath}{path to save files returened by the function.}
}
\description{
This Function is used to plot Gene ontolgy summary in the data of the accession/s.
}
\author{
Mohmed Soudy \email{Mohamed.soudy@57357.com} and Ali Mostafa \email{ali.mo.anwar@std.agr.cu.edu.eg}
}
