% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotGoterms.R
\name{PlotGOBiological}
\alias{PlotGOBiological}
\title{Connect and parse UniProt information.}
\usage{
PlotGOBiological(GOObj, Top = 10, directorypath = NULL)
}
\arguments{
\item{GOObj}{Dataframe returned from UniprotR Function "GetProteinGOInfo"}

\item{Top}{Number of molecular functions to be visualized}

\item{directorypath}{path to save Output plot.}
}
\description{
This Function is used to plot Biological process of proteins.
}
\author{
Mohmed Soudy \email{Mohamed.soudy@57357.com} and Ali Mostafa \email{ali.mo.anwar@std.agr.cu.edu.eg}
}
