% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotGoInfo.R
\name{PlotGoInfo}
\alias{PlotGoInfo}
\title{Connect and parse UniProt information.}
\usage{
PlotGoInfo(GOObj , directorypath = NULL)
}
\arguments{
\item{GOObj}{Dataframe.}

\item{directorypath}{path to save excel file containig results returened by the function ( default = NA ).}
}
\description{
This Function is used to plot data retrieved from UniprotR Function "Goparse".
}
\author{
Mohmed Soudy \email{Mohamed.soudy@57357.com} and Ali Mostafa \email{ali.mo.anwar@std.agr.cu.edu.eg}
}
