% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotGoterms.R
\name{PlotGoterms}
\alias{PlotGoterms}
\title{Connect and parse UniProt information.}
\usage{
PlotGoterms(GOObj , directorypath = NULL)
}
\arguments{
\item{GOObj}{Dataframe.}

\item{directorypath}{path to save plot returned by function ( default = NA ).}
}
\description{
This Function is used to plot data retrieved from UniprotR Function "GetProteinGOInfo".
}
\author{
Mohmed Soudy \email{Mohamed.soudy@57357.com} and Ali Mostafa \email{ali.mo.anwar@std.agr.cu.edu.eg}
}
