% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotPhyscochemical.R
\name{PlotPhysicochemical}
\alias{PlotPhysicochemical}
\title{Connect and parse UniProt information.}
\usage{
PlotPhysicochemical(SeqDataObjPath , directorypath = NULL)
}
\arguments{
\item{SeqDataObjPath}{Dataframe returned from GetSequence function.}

\item{directorypath}{Path to save Physcochemical properties plot.}
}
\description{
This function can be used to get a list of UniProt Accession/s from a csv file.
}
\author{
Mohmed Soudy \email{Mohamed.soudy@57357.com} and Ali Mostafa \email{ali.mo.anwar@std.agr.cu.edu.eg}
}
