% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotproteinExist.R
\name{PlotproteinExist}
\alias{PlotproteinExist}
\title{Connect and parse UniProt information.}
\usage{
PlotproteinExist(ProteinDataObject,directorypath = NULL)
}
\arguments{
\item{ProteinDataObject}{input a Dataframe returned from GetMiscellaneous function}

\item{directorypath}{path to save files returened by the function.}
}
\description{
This Function is used to plot protein status in the data of the accession/s.
}
\author{
Mohmed Soudy \email{Mohamed.soudy@57357.com} and Ali Mostafa \email{ali.mo.anwar@std.agr.cu.edu.eg}
}
