% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_GO.R
\name{PlotProteinGO_molc}
\alias{PlotProteinGO_molc}
\title{Connect and parse UniProt information.}
\usage{
PlotProteinGO_molc(GO_df , dir_path = NA)
}
\arguments{
\item{GO_df}{Dataframe.}

\item{dir_path}{path to save files returened by the function ( default = NA ).}
}
\description{
This Function is used to plot molecular functions data retrieved from UniprotR Function "GetProteinGOInfo".
}
\note{
if no dir_path was given ( default = NA ) the function will only view the plot and will not save it
}
\author{
Mohmed Soudy \email{Mohamed.soudy@57357.com} and Ali Mostafa \email{ali.mo.anwar@std.agr.cu.edu.eg}
}
