% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotProteinTaxa.R
\name{PlotProteinsLoc}
\alias{PlotProteinsLoc}
\title{Connect and parse UniProt information.}
\usage{
PlotProteinsLoc(ProteinsData, directorypath = NULL)
}
\arguments{
\item{ProteinsData}{input a Dataframe of proteins as rownames.}

\item{directorypath}{path to save excel file containig results returened by the function.}
}
\description{
This Function is used to plot different locations where accession/s can be found in.
}
\author{
Mohmed Soudy and Ali Mostafa
}
