% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotProteinTaxa.R
\name{PlotProteinTaxa}
\alias{PlotProteinTaxa}
\title{Connect and parse UniProt information.}
\usage{
PlotProteinTaxa(ProteinDataObject , directorypath = NULL)
}
\arguments{
\item{ProteinDataObject}{input a Dataframe of proteins as rownames.}

\item{directorypath}{path to save files returened by the function.}
}
\description{
This Function is used to plot different taxas found of the accessions.
}
\author{
Mohmed Soudy \email{Mohamed.soudy@57357.com} and Ali Mostafa \email{ali.mo.anwar@std.agr.cu.edu.eg}
}
