% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotSummaryInfo.R
\name{PlotSummaryInfo}
\alias{PlotSummaryInfo}
\title{Connect and parse UniProt information.}
\usage{
PlotSummaryInfo (ProteinDataObject,directorypath = NULL)
}
\arguments{
\item{ProteinDataObject}{input a Dataframe returned from GetNamesTaxa function}

\item{directorypath}{path to save excel file containig results returened by the function.}
}
\description{
This Function is used to plot location's frequency in the data of the accession/s in the chromosomes.
}
\author{
Mohmed Soudy and Ali Mostafa
}
