% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ProcessAcc.R
\name{ProcessAcc}
\alias{ProcessAcc}
\title{Connect and parse UniProt information.}
\usage{
ProcessAcc(Accessions)
}
\arguments{
\item{Accessions}{acession list returened from GetAccession function}
}
\description{
This Function is used to check validty of input accessions the data of the accession/s.
}
\author{
Mohmed Soudy \email{Mohamed.soudy@57357.com} and Ali Mostafa \email{ali.mo.anwar@std.agr.cu.edu.eg}
}
