% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UniprotR.r
\docType{package}
\name{UniprotR}
\alias{UniprotR}
\alias{UniprotR-package}
\title{UniprotR: A package for retrieving proteins information from UniProt.}
\description{
UniprotR provides different functions to retrieve proteins information provided by the UniProt.
UniprotR could be divided into two categories one for retrieving and one for visualizing the data.
}
\section{The first category for retrieving data}{

____________________________________________
ConstructLocTree

GetCross_references_Information

GetExpression

GetFamily_Domains

GetGeneral_Information

GetMiscellaneous

GetNamesTaxa

GetPathology_Biotech

GetProteinFunction

GetProteinGOInfo

GetProteinInteractions

GetProteoAnnotator

GetPTM_Processsing

GetPublication

GetSequences

GetSubcellular_location

GetTaxaidentifier

GetTaxonomic_lineage


each function has a description that could be reached by >>> ?Function_Name
}

\section{The second category for visualizing data}{

_______________________________________________

PlotProteinTaxa

PlotChromosomeInfo

PlotGoInfo

plotGoannotation

PlotProteinGO_bio

PlotProteinGO_cel

PlotProteinGO_molc

PlotproteinScore

PlotproteinStatus

PlotproteinExist

each function has a description that could be reached by  >>> ?Function_Name
}

\author{
Mohmed Soudy \email{Mohamed.soudy@57357.com} and Ali Mostafa \email{ali.mo.anwar@std.agr.cu.edu.eg}
}
