% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GETSeqFastaUniprot.R
\name{GETSeqFastaUniprot}
\alias{GETSeqFastaUniprot}
\title{Connect and parse UniProt information.}
\usage{
GETSeqFastaUniprot(Accessions, FileName = NULL)
}
\arguments{
\item{Accessions}{list of Uniprot accessions}

\item{FileName}{OUtput file name}
}
\description{
This Function is used to download sequences of given protein list in a fasta format
}
\author{
Mohmed Soudy \email{Mohamed.soudy@57357.com} and Ali Mostafa \email{ali.mo.anwar@std.agr.cu.edu.eg}
}
