% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotProteinTaxa.R
\name{PlotProteinTaxa}
\alias{PlotProteinTaxa}
\title{Connect and parse UniProt information.}
\usage{
PlotProteinTaxa(ProteinsData , directorypath = NULL)
}
\arguments{
\item{ProteinsData}{input a Dataframe of proteins as rownames.}

\item{directorypath}{path to save excel file containig results returened by the function.}
}
\description{
This Function is used to plot different taxas found of the accessions.
}
\author{
Mohmed Soudy and Ali Mostafa
}
